Expose number of near and far pointers, factor out state machine.
