+++ Assets/MixedRealityToolkit.Services/InputSystem/GazePointerStateMachine.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a609eb6100c33af4b8727a4d7b6e4fdc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
